Move VLC native download logic into a proper Gradle task class
Replaces the buildscript {} + inline helper functions in
flixelgdx-video-lwjgl3 with a DownloadVlcNativesTask class in build-logic.
commons-compress and xz are now implementation deps of the included build,
eliminating the meta-classpath buildscript block and the deprecated
SevenZFile(File) constructor (now uses SevenZFile.Builder). FileSystemOperations
and ArchiveOperations are injected, replacing the deprecated project.copy /
project.delete calls in the task action.
